machine_learning/lang_chain/requirements.txt

requests
python-dotenv
langchain-ollama
langchain[groq]